--- a/sbx/sac/sac.py
+++ b/sbx/sac/sac.py
@@ -1,7 +1,6 @@
 from functools import partial
 from typing import Any, ClassVar, Dict, Optional, Tuple, Type, Union
 
-import flax
 import flax.linen as nn
 import jax
 import jax.numpy as jnp
@@ -9,6 +8,7 @@ import numpy as np
 import optax
 from flax.training.train_state import TrainState
 from gymnasium import spaces
+from jax.typing import ArrayLike
 from stable_baselines3.common.buffers import ReplayBuffer
 from stable_baselines3.common.noise import ActionNoise
 from stable_baselines3.common.type_aliases import GymEnv, MaybeCallback, Schedule
@@ -182,11 +182,6 @@ class SAC(OffPolicyAlgorithmJax):
     def train(self, batch_size, gradient_steps):
         # Sample all at once for efficiency (so we can jit the for loop)
         data = self.replay_buffer.sample(batch_size * gradient_steps, env=self._vec_normalize_env)
-        # Pre-compute the indices where we need to update the actor
-        # This is a hack in order to jit the train loop
-        # It will compile once per value of policy_delay_indices
-        policy_delay_indices = {i: True for i in range(gradient_steps) if ((self._n_updates + i + 1) % self.policy_delay) == 0}
-        policy_delay_indices = flax.core.FrozenDict(policy_delay_indices)
 
         if isinstance(data.observations, dict):
             keys = list(self.observation_space.keys())
@@ -217,7 +212,8 @@ class SAC(OffPolicyAlgorithmJax):
             self.target_entropy,
             gradient_steps,
             data,
-            policy_delay_indices,
+            self.policy_delay,
+            self._n_updates % self.policy_delay,
             self.policy.qf_state,
             self.policy.actor_state,
             self.ent_coef_state,
@@ -236,11 +232,11 @@ class SAC(OffPolicyAlgorithmJax):
         actor_state: TrainState,
         qf_state: RLTrainState,
         ent_coef_state: TrainState,
-        observations: np.ndarray,
-        actions: np.ndarray,
-        next_observations: np.ndarray,
-        rewards: np.ndarray,
-        dones: np.ndarray,
+        observations: jax.Array,
+        actions: jax.Array,
+        next_observations: jax.Array,
+        rewards: jax.Array,
+        dones: jax.Array,
         key: jax.random.KeyArray,
     ):
         key, noise_key, dropout_key_target, dropout_key_current = jax.random.split(key, 4)
@@ -284,7 +280,7 @@ class SAC(OffPolicyAlgorithmJax):
         actor_state: RLTrainState,
         qf_state: RLTrainState,
         ent_coef_state: TrainState,
-        observations: np.ndarray,
+        observations: jax.Array,
         key: jax.random.KeyArray,
     ):
         key, dropout_key, noise_key = jax.random.split(key, 3)
@@ -319,7 +315,7 @@ class SAC(OffPolicyAlgorithmJax):
 
     @staticmethod
     @jax.jit
-    def update_temperature(target_entropy: np.ndarray, ent_coef_state: TrainState, entropy: float):
+    def update_temperature(target_entropy: ArrayLike, ent_coef_state: TrainState, entropy: float):
         def temperature_loss(temp_params):
             ent_coef_value = ent_coef_state.apply_fn({"params": temp_params})
             ent_coef_loss = ent_coef_value * (entropy - target_entropy).mean()
@@ -331,62 +327,112 @@ class SAC(OffPolicyAlgorithmJax):
         return ent_coef_state, ent_coef_loss
 
     @classmethod
-    @partial(jax.jit, static_argnames=["cls", "gradient_steps"])
+    def update_actor_and_temperature(
+        cls,
+        actor_state: RLTrainState,
+        qf_state: RLTrainState,
+        ent_coef_state: TrainState,
+        observations: jax.Array,
+        target_entropy: ArrayLike,
+        key: jax.random.KeyArray,
+    ):
+        (actor_state, qf_state, actor_loss_value, key, entropy) = cls.update_actor(
+            actor_state,
+            qf_state,
+            ent_coef_state,
+            observations,
+            key,
+        )
+        ent_coef_state, ent_coef_loss_value = cls.update_temperature(target_entropy, ent_coef_state, entropy)
+        return actor_state, qf_state, ent_coef_state, actor_loss_value, ent_coef_loss_value, key
+
+    @classmethod
+    @partial(jax.jit, static_argnames=["cls", "gradient_steps", "policy_delay_interval", "policy_delay_offset"])
     def _train(
         cls,
         gamma: float,
         tau: float,
-        target_entropy: np.ndarray,
+        target_entropy: ArrayLike,
         gradient_steps: int,
         data: ReplayBufferSamplesNp,
-        policy_delay_indices: flax.core.FrozenDict,
+        policy_delay_interval: int,
+        policy_delay_offset: int,
         qf_state: RLTrainState,
         actor_state: TrainState,
         ent_coef_state: TrainState,
-        key,
+        key: jax.random.KeyArray,
     ):
-        actor_loss_value = jnp.array(0)
+        assert data.observations.shape[0] % gradient_steps == 0
+        batch_size = data.observations.shape[0] // gradient_steps
 
-        for i in range(gradient_steps):
+        carry = {
+            "actor_state": actor_state,
+            "qf_state": qf_state,
+            "ent_coef_state": ent_coef_state,
+            "key": key,
+            "info": {
+                "actor_loss": jnp.array(0.0),
+                "qf_loss": jnp.array(0.0),
+                "ent_coef_loss": jnp.array(0.0),
+            },
+        }
 
-            def slice(x, step=i):
-                assert x.shape[0] % gradient_steps == 0
-                batch_size = x.shape[0] // gradient_steps
-                return x[batch_size * step : batch_size * (step + 1)]
-
+        def one_update(i: int, carry: Dict[str, Any]) -> Dict[str, Any]:
+            actor_state = carry["actor_state"]
+            qf_state = carry["qf_state"]
+            ent_coef_state = carry["ent_coef_state"]
+            key = carry["key"]
+            info = carry["info"]
+            batch_obs = jax.lax.dynamic_slice_in_dim(data.observations, i, batch_size)
+            batch_act = jax.lax.dynamic_slice_in_dim(data.actions, i, batch_size)
+            batch_next_obs = jax.lax.dynamic_slice_in_dim(data.next_observations, i, batch_size)
+            batch_rew = jax.lax.dynamic_slice_in_dim(data.rewards, i, batch_size)
+            batch_done = jax.lax.dynamic_slice_in_dim(data.dones, i, batch_size)
             (
                 qf_state,
                 (qf_loss_value, ent_coef_value),
                 key,
-            ) = SAC.update_critic(
+            ) = cls.update_critic(
                 gamma,
                 actor_state,
                 qf_state,
                 ent_coef_state,
-                slice(data.observations),
-                slice(data.actions),
-                slice(data.next_observations),
-                slice(data.rewards),
-                slice(data.dones),
+                batch_obs,
+                batch_act,
+                batch_next_obs,
+                batch_rew,
+                batch_done,
                 key,
             )
-            qf_state = SAC.soft_update(tau, qf_state)
+            qf_state = cls.soft_update(tau, qf_state)
 
-            # hack to be able to jit (n_updates % policy_delay == 0)
-            if i in policy_delay_indices:
-                (actor_state, qf_state, actor_loss_value, key, entropy) = cls.update_actor(
-                    actor_state,
-                    qf_state,
-                    ent_coef_state,
-                    slice(data.observations),
-                    key,
-                )
-                ent_coef_state, _ = SAC.update_temperature(target_entropy, ent_coef_state, entropy)
+            (actor_state, qf_state, ent_coef_state, actor_loss_value, ent_coef_loss_value, key) = jax.lax.cond(
+                (policy_delay_offset + i) % policy_delay_interval == 0,
+                cls.update_actor_and_temperature,
+                lambda *_: (actor_state, qf_state, ent_coef_state, info["actor_loss"], info["ent_coef_loss"], key),
+                actor_state,
+                qf_state,
+                ent_coef_state,
+                batch_obs,
+                target_entropy,
+                key,
+            )
+            info = {"actor_loss": actor_loss_value, "qf_loss": qf_loss_value, "ent_coef_loss": ent_coef_loss_value}
+
+            return {
+                "actor_state": actor_state,
+                "qf_state": qf_state,
+                "ent_coef_state": ent_coef_state,
+                "key": key,
+                "info": info,
+            }
+
+        update_carry = jax.lax.fori_loop(0, gradient_steps, one_update, carry)
 
         return (
-            qf_state,
-            actor_state,
-            ent_coef_state,
+            update_carry["qf_state"],
+            update_carry["actor_state"],
+            update_carry["ent_coef_state"],
             key,
-            (actor_loss_value, qf_loss_value, ent_coef_value),
+            (update_carry["info"]["actor_loss"], update_carry["info"]["qf_loss"], update_carry["info"]["ent_coef_loss"]),
         )