--- a/sbx/td3/td3.py
+++ b/sbx/td3/td3.py
@@ -1,7 +1,6 @@
 from functools import partial
 from typing import Any, ClassVar, Dict, Optional, Tuple, Type, Union
 
-import flax
 import jax
 import jax.numpy as jnp
 import numpy as np
@@ -124,11 +123,6 @@ class TD3(OffPolicyAlgorithmJax):
     def train(self, batch_size, gradient_steps):
         # Sample all at once for efficiency (so we can jit the for loop)
         data = self.replay_buffer.sample(batch_size * gradient_steps, env=self._vec_normalize_env)
-        # Pre-compute the indices where we need to update the actor
-        # This is a hack in order to jit the train loop
-        # It will compile once per value of policy_delay_indices
-        policy_delay_indices = {i: True for i in range(gradient_steps) if ((self._n_updates + i + 1) % self.policy_delay) == 0}
-        policy_delay_indices = flax.core.FrozenDict(policy_delay_indices)
 
         if isinstance(data.observations, dict):
             keys = list(self.observation_space.keys())
@@ -157,7 +151,8 @@ class TD3(OffPolicyAlgorithmJax):
             self.tau,
             gradient_steps,
             data,
-            policy_delay_indices,
+            self.policy_delay,
+            self._n_updates % self.policy_delay,
             self.target_policy_noise,
             self.target_noise_clip,
             self.policy.qf_state,
@@ -175,11 +170,11 @@ class TD3(OffPolicyAlgorithmJax):
         gamma: float,
         actor_state: RLTrainState,
         qf_state: RLTrainState,
-        observations: np.ndarray,
-        actions: np.ndarray,
-        next_observations: np.ndarray,
-        rewards: np.ndarray,
-        dones: np.ndarray,
+        observations: jax.Array,
+        actions: jax.Array,
+        next_observations: jax.Array,
+        rewards: jax.Array,
+        dones: jax.Array,
         target_policy_noise: float,
         target_noise_clip: float,
         key: jax.random.KeyArray,
@@ -222,7 +217,7 @@ class TD3(OffPolicyAlgorithmJax):
     def update_actor(
         actor_state: RLTrainState,
         qf_state: RLTrainState,
-        observations: np.ndarray,
+        observations: jax.Array,
         key: jax.random.KeyArray,
     ):
         key, dropout_key = jax.random.split(key, 2)
@@ -256,60 +251,86 @@ class TD3(OffPolicyAlgorithmJax):
         return qf_state, actor_state
 
     @classmethod
-    @partial(jax.jit, static_argnames=["cls", "gradient_steps"])
+    @partial(jax.jit, static_argnames=["cls", "gradient_steps", "policy_delay_interval", "policy_delay_offset"])
     def _train(
         cls,
         gamma: float,
         tau: float,
         gradient_steps: int,
         data: ReplayBufferSamplesNp,
-        policy_delay_indices: flax.core.FrozenDict,
+        policy_delay_interval: int,
+        policy_delay_offset: int,
         target_policy_noise: float,
         target_noise_clip: float,
         qf_state: RLTrainState,
         actor_state: RLTrainState,
-        key,
+        key: jax.random.KeyArray,
     ):
-        actor_loss_value = jnp.array(0)
+        assert data.observations.shape[0] % gradient_steps == 0
+        batch_size = data.observations.shape[0] // gradient_steps
 
-        for i in range(gradient_steps):
+        carry = {
+            "actor_state": actor_state,
+            "qf_state": qf_state,
+            "key": key,
+            "info": {
+                "actor_loss": jnp.array(0.0),
+                "qf_loss": jnp.array(0.0),
+            },
+        }
 
-            def slice(x, step=i):
-                assert x.shape[0] % gradient_steps == 0
-                batch_size = x.shape[0] // gradient_steps
-                return x[batch_size * step : batch_size * (step + 1)]
-
+        def one_update(i: int, carry: Dict[str, Any]) -> Dict[str, Any]:
+            actor_state = carry["actor_state"]
+            qf_state = carry["qf_state"]
+            key = carry["key"]
+            info = carry["info"]
+            batch_obs = jax.lax.dynamic_slice_in_dim(data.observations, i, batch_size)
+            batch_act = jax.lax.dynamic_slice_in_dim(data.actions, i, batch_size)
+            batch_next_obs = jax.lax.dynamic_slice_in_dim(data.next_observations, i, batch_size)
+            batch_rew = jax.lax.dynamic_slice_in_dim(data.rewards, i, batch_size)
+            batch_done = jax.lax.dynamic_slice_in_dim(data.dones, i, batch_size)
             (
                 qf_state,
                 qf_loss_value,
                 key,
-            ) = TD3.update_critic(
+            ) = cls.update_critic(
                 gamma,
                 actor_state,
                 qf_state,
-                slice(data.observations),
-                slice(data.actions),
-                slice(data.next_observations),
-                slice(data.rewards),
-                slice(data.dones),
+                batch_obs,
+                batch_act,
+                batch_next_obs,
+                batch_rew,
+                batch_done,
                 target_policy_noise,
                 target_noise_clip,
                 key,
             )
-            qf_state, actor_state = TD3.soft_update(tau, qf_state, actor_state)
+            qf_state, actor_state = cls.soft_update(tau, qf_state, actor_state)
 
-            # hack to be able to jit (n_updates % policy_delay == 0)
-            if i in policy_delay_indices:
-                (actor_state, qf_state, actor_loss_value, key) = cls.update_actor(
-                    actor_state,
-                    qf_state,
-                    slice(data.observations),
-                    key,
-                )
+            (actor_state, qf_state, actor_loss_value, key) = jax.lax.cond(
+                (policy_delay_offset + i) % policy_delay_interval == 0,
+                cls.update_actor,
+                lambda *_: (actor_state, qf_state, info["actor_loss"], key),
+                actor_state,
+                qf_state,
+                batch_obs,
+                key,
+            )
+            info = {"actor_loss": actor_loss_value, "qf_loss": qf_loss_value}
+
+            return {
+                "actor_state": actor_state,
+                "qf_state": qf_state,
+                "key": key,
+                "info": info,
+            }
+
+        update_carry = jax.lax.fori_loop(0, gradient_steps, one_update, carry)
 
         return (
-            qf_state,
-            actor_state,
+            update_carry["qf_state"],
+            update_carry["actor_state"],
             key,
-            (actor_loss_value, qf_loss_value),
+            (update_carry["info"]["actor_loss"], update_carry["info"]["qf_loss"]),
         )