… README.md
… sbx/tqc/tqc.py
… sbx/version.txt
… setup.py