<!--- Provide a general summary of your changes in the Title above -->

## Description
closes #14 

Addresses #14 by replacing the python loop in `_train()` with a `jax.lax.fori_loop` construct for SAC, TD3, and TQC.
This avoids explicitly unrolling the loop during compile time, which significantly reduces compile times and the size of the compiled program for large `gradient_steps` values.

Since performance was a concern in https://github.com/araffin/sbx/issues/14#issuecomment-1687684993, I did some benchmarking. To not exacerbate the influence of the compile time, I trained each agent for 1 million steps. I used different values for `train_freq` (with `gradient_steps=-1`) to highlight the differences between the implementations.
I used the following script for benchmarking.
```python
import argparse
import time

from sbx import SAC, TD3, TQC


parser = argparse.ArgumentParser()
parser.add_argument("algo", type=str)
parser.add_argument("train_freq", type=int)
parser.add_argument("--steps", type=int, default=1_000_000)
args = parser.parse_args()

algo_class = eval(args.algo.upper())

start = time.time()
algo = algo_class(
    "MlpPolicy",
    "Walker2d-v4",
    train_freq=args.train_freq,
    gradient_steps=-1,
    tensorboard_log=f"tensorboard/{args.algo}_{args.train_freq}_{args.steps}",
)
algo.learn(args.steps)
print(f"{args.algo.upper()}, train_freq: {args.train_freq}, steps: {args.steps}: {time.time() - start:.2f}s")
```

I got the following results on Ubuntu 22.04 with an NVIDIA RTX A6000 GPU.

SAC:

| train_freq and gradient_steps | Execution time in seconds (original) | Execution time in seconds (PR) | Relative speedup (t<sub>orig</sub> / t<sub>pr</sub>) |
|-------------------------------|--------------------------------------|--------------------------------|------------------------------------------------------| 
| 1                             | 2716                                 | 2713                           | 100.1 %                                              |
| 4                             | 1621                                 | 1659                           | 97.7 %                                               |
| 16                            | 1498                                 | 1490                           | 100.5 %                                              |
| 64                            | 1784                                 | 1360                           | 130.2 %                                              |
| 128                           | 2587                                 | 1256                           | 206.0 %                                              |

TD3:

| train_freq and gradient_steps | Execution time in seconds (original) | Execution time in seconds (PR) | Relative speedup (t<sub>orig</sub> / t<sub>pr</sub>) |
|-------------------------------|--------------------------------------|--------------------------------|------------------------------------------------------|
| 1                             | 2281                                 | 2270                           | 100.5 %                                              |
| 4                             | 1315                                 | 1404                           | 93.7 %                                               |
| 16                            | 1094                                 | 1070                           | 102.2 %                                              |
| 64                            | 1103                                 | 950                            | 116.1 %                                              |
| 128                           | 1311                                 | 939                            | 139.6 %                                              |

TQC:

| train_freq and gradient_steps | Execution time in seconds (original) | Execution time in seconds (PR) | Relative speedup (t<sub>orig</sub> / t<sub>pr</sub>) |
|-------------------------------|--------------------------------------|--------------------------------|------------------------------------------------------|
| 1                             | 2995                                 | 2936                           | 102.0 %                                              |
| 4                             | 2070                                 | 1926                           | 107.5 %                                              |
| 16                            | 2035                                 | 1721                           | 118.2 %                                              |
| 64                            | 2090                                 | 1443                           | 144.8 %                                              |
| 128                           | 3359                                 | 1418                           | 236.9 %                                              |

For small `gradient_steps` values, the implementation is on par with the original implementation for all three algorithms. For large values, it leads to significant speed-ups due to faster compilation and smaller program sizes.
I also tried running the original implementation with `gradient_steps=1000` but it ran out-of-memory during compilation on my 64GB RAM machine.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
<!--- You can use the syntax `closes #100` if this solves the issue #100 -->
- [x] I have raised an issue to propose this change ([required](https://github.com/DLR-RM/stable-baselines3/blob/master/CONTRIBUTING.md) for new features and bug fixes)

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation (update in the documentation)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I've read the [CONTRIBUTION](https://github.com/DLR-RM/stable-baselines3/blob/master/CONTRIBUTING.md) guide (**required**)
- [ ] I have updated the changelog accordingly (**required**).
       I could not find the changelog (there is no `docs/misc/changelog.rst`)
- [ ] My change requires a change to the documentation.
- [ ] I have updated the tests accordingly (*required for a bug fix or a new feature*).
- [ ] I have updated the documentation accordingly.
- [x] I have reformatted the code using `make format` (**required**)
- [x] I have checked the codestyle using `make check-codestyle` and `make lint` (**required**)
- [x] I have ensured `make pytest` and `make type` both pass. (**required**)
- [ ] I have checked that the documentation builds using `make doc` (**required**)

Note: You can run most of the checks using `make commit-checks`.

Note: we are using a maximum length of 127 characters per line

<!--- This Template is an edited version of the one from https://github.com/evilsocket/pwnagotchi/ -->