@araffin 2023-12-13
Hello,
thanks a lot for the PR =)
I'll try to look into it today.

I assume you also check that the performance was the same with this change?