@jan1854 2023-12-15
> I assume you also check that the performance was the same with this change?

I benchmarked the performance of the three algorithms (after the bugfix) on Walker2D-v4 for 10 random seeds. I used the default hyperparameter (the `train_freq=1` case of the computation time benchmark above). Here are the results (mean and standard deviation across the seeds).

<img src='https://github.com/araffin/sbx/assets/33448112/56980c65-e0fa-4b2f-96e2-a756aa12f650' width='260'> 
<img src='https://github.com/araffin/sbx/assets/33448112/dcb36226-cd1f-4044-9325-5a3d83cc260b' width='260'> 
<img src='https://github.com/araffin/sbx/assets/33448112/67b17de0-3a62-4245-9d2f-904ae1789a62' width='260'>

In all cases, the average performance is roughly the same for the original implementation and the PR. The standard deviation differs between the implementations for SAC and TD3 but I assume this is due to noise and that the difference would vanish for more seeds.