@araffin 2024-01-11
Hello,
I'm back =)
thanks for the benchmark, I will try to do one additional soon with `train_freq=gradient_steps=12`.