@araffin 2024-01-29
After some local tests:
- there seems to be a bug with newer jax versions (I get NaN when I try to unroll on CUDA)
- the performance drop also depends on jax/cuda version, the compilation time too (latest jax version, 0.4.23 is slower to compile, yield NaNs when trying to unroll but is the fastest, no perf drop w.r.t. master)

Tested jax versions: 
- `jaxlib==0.4.12+cuda11.cudnn86` (work with unroll but slower when using for i loo) 
- `jaxlib==0.4.23+cuda11.cudnn86` (NaN with unroll, faster than master when using for i loop, couldn't test master though)

I would be happy if you could help isolate the NaN bug so we can send a minimal working example to Jax team.