@jan1854 2024-02-13
> After some local tests:
> * there seems to be a bug with newer jax versions (I get NaN when I try to unroll on CUDA)
> * the performance drop also depends on jax/cuda version, the compilation time too (latest jax version, 0.4.23 is slower to compile, yield NaNs when trying to unroll but is the fastest, no perf drop w.r.t. master)
> 
> 
> Tested jax versions:
>  * `jaxlib==0.4.12+cuda11.cudnn86` (work with unroll but slower when using for i loo)
>  * `jaxlib==0.4.23+cuda11.cudnn86` (NaN with unroll, faster than master when using for i loop, couldn't test master though)

Hi @araffin, I was able to reproduce the NaN problem on my machine for `jaxlib==0.4.23+cuda11.cudnn86`. However, I noticed that the problem vanishes for jaxlib version 0.4.24. It also seems to be independent of whether the CUDA 11 or CUDA 12 version is used (works for both `jaxlib==0.4.24+cuda11.cudnn86` or `jaxlib==0.4.24+cuda12.cudnn89`). So, I guess that the Jax people fixed the issue in the latest jaxlib version, but I could not find anything about it in the [changelogs](https://github.com/google/jax/blob/main/CHANGELOG.md).

I will do some benchmarking of the performance with different values for the unroll parameter to figure out if it helps. I'll come back to you as soon as I have the data.