@araffin 2024-02-14
Thanks, I can confirm that the bug is gone with newest version of jax.
After some testing, it also seems that keeping the default `unroll` value is the best for speed.