@jan1854 2024-02-15
I can confirm that observation. In my tests, the runs without loop unrolling were almost always faster (in terms of the `time/fps` metric). I also compared against the old implementation from before the PR. Previously, you reported that `time/fps` was slightly higher for the old implementation. In my experiments, this does not seem to be the case anymore. The new implementation was always faster, so perhaps the new version of Jax also improved the efficiency of the `fori_loop` construct in the meantime.

The results seem to be consistent across all three algorithms.